[OPS-2598] Fixed Problems from Pull Request
